scheme change, jwt auth

+++ src/user/user.controller.ts
@@ -3,31 +3,49 @@ import {
   Controller,
   Delete,
   Get,
+  NotFoundException,
   Param,
+  ParseIntPipe,
   ParseUUIDPipe,
   Post,
   UseGuards,
 } from '@nestjs/common';
 import { UserService } from './user.service';
 import { AuthGuard } from 'src/auth/auth.guard';
-import { ApiTags } from '@nestjs/swagger';
+import { ApiNotFoundResponse, ApiOkResponse, ApiTags } from '@nestjs/swagger';
+import { User } from '@prisma/client';
+import { UserDto } from './dto/user.dto';
 @ApiTags('User')
 @Controller('user')
 export class UserController {
   constructor(private readonly userService: UserService) {}
 
-  @Post()
-  createUser(@Body() dto) {
-    return this.userService.save(dto);
+  // @Post()
+  // createUser(@Body() dto) {
+  //   return this.userService.save(dto);
+  // }
+  @ApiOkResponse({
+    type: UserDto,
+  })
+  @ApiNotFoundResponse({
+    description: 'User not found',
+  })
+  @Get(':idOrEmail')
+  async findOneUser(@Param('idOrEmail') idOrEmail: string): Promise<UserDto> {
+    return await this.userService.findOne(idOrEmail);
   }
 
-  @Get(':idOrEmail')
-  findOneUser(@Param('idOrEmail') idOrEmail: string) {
-    return this.userService.findOne(idOrEmail);
-  }
+  @ApiOkResponse({
+    type: UserDto,
+  })
+  @ApiNotFoundResponse({
+    description: 'User not found',
+  })
   @UseGuards(AuthGuard)
   @Delete(':id')
-  deleteUser(@Param('id', ParseUUIDPipe) id: string) {
-    return this.userService.delete(id);
+  async deleteUser(@Param('id', ParseIntPipe) id: number): Promise<UserDto> {
+    const existingUser = await this.userService.findOne(id);
+    if (!existingUser) throw new NotFoundException('User not found');
+    return await this.userService.delete(id);
   }
 }
